pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lackey"
version = "0.1.0"
description = "An unattended coding agent inspired by Stripe's Minions"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pydantic-ai-slim[anthropic]>=0.2",
    "boto3>=1.34",
    "PyJWT[crypto]>=2.0",
]

[project.scripts]
lackey = "lackey.run:main"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "N",     # pep8-naming
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "SIM",   # flake8-simplify
    "T20",   # flake8-print
    "RUF",   # ruff-specific rules
]

[tool.ruff.lint.isort]
known-first-party = ["lackey"]
